Implementacion de validaciones para CRUD Rooms y Client

+++ Backend/src/application/use-cases/client/deleteClient.ts
@@ -5,6 +5,12 @@ export class DeleteClient {
   constructor(private repository: IClientRepository) {}
 
   async delete(id: string): Promise<void> {
+    const existClient = await this.repository.existClient(new clientId(id));
+
+    if (!existClient) {
+      throw new Error("Cliente no encontrado.");
+    }
+
     await this.repository.deleteClient(new clientId(id));
   }
 }
